+++ src/Parsable.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Jenky\Hermes;
+
+interface Parsable
+{
+    /**
+     * Parse the response body to native type.
+     *
+     * @return void
+     */
+    public function parse();
+}