+++ CHANGELOG.md
@@ -16,14 +16,14 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ### Fixed -->
 
-## [Unreleased]
+## [UNRELEASED]
 
 ### Added
 - Add some syntax sugar methods to interact with HTTP response message
 
 ### Changed
 - Use splat operator for `array_merge_recursive_unique` instead of `func_get_args()`
-- Use `Jenky\Hermes\Transformable` interface for response handler that needs to transform response body to native type
+- Use `Jenky\Hermes\Transformable` interface for response handler that needs to parse response body to native type
 
 ## [1.1.1](https://github.com/jenky/hermes/compare/1.1.0...1.1.1) - 2020-02-25